% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScan.R
\name{BayeScan}
\alias{BayeScan}
\title{Create BayeScan object}
\usage{
BayeScan(opts, data, results)
}
\arguments{
\item{opts}{\code{BayeScanOpts} object with parameters used to run BayeScan .}

\item{data}{\code{BayeScanData} object with input data used for analysis.}

\item{results}{\code{BayeScanResults} object with results from analysis.}
}
\description{
This function creates a new \code{BayeScan} object.
}
\seealso{
\code{\link{BayeScan-class}}, \code{\link{BayeScanData}}, \code{\link{BayeScanData}}, \code{\link{BayeScanResults}}.
}

