% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanData.R
\name{BayeScanData}
\alias{BayeScanData}
\title{Create BayeScanData object}
\usage{
BayeScanData(matrix, primers, populations)
}
\arguments{
\item{matrix}{\code{matrix} with binary loci data. Each row is for a different sample; each column is for a different loci.}

\item{primers}{\code{character} names of the selective primers used for the AFLPs followed by the location on the gel.}

\item{populations}{\code{character} name of populations that each sample is from.}
}
\description{
This function creates a new \code{BayeScanData} object.
}
\seealso{
\code{\link{BayeScanData-class}}.
}

