% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\name{BayeScanResults}
\alias{BayeScanResults}
\title{Create BayeScanResults object}
\usage{
BayeScanResults(fst, mcmc, acceptance.rate, verification)
}
\value{
\code{\link{BayeScanResults}}.
}
\description{
This function creates a new \code{BayeScanResults} object.
}
\seealso{
\code{\link{BayeScanResults-class}}.
}

