% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractResults.R
\name{extractResults}
\alias{extractResults}
\title{Extract results from RapSolved object}
\usage{
extractResults(x)
}
\arguments{
\item{x}{\code{RapSolved} object.}
}
\description{
This function extracts results from a RapSolved object
}

