% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.table.R
\name{format.table}
\alias{format.table}
\title{Format table for knitr}
\usage{
\method{format}{table}(x)
}
\arguments{
\item{x}{\code{data.frame}}
}
\value{
\code{data.frame}
}
\description{
This function formats tables for rendering in scientific reports.
}

