% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R, R/BayeScan.R
\name{loci.subset}
\alias{loci.subset}
\alias{loci.subset.BayeScan}
\alias{loci.subset.BayeScanData}
\title{Subset loci}
\usage{
loci.subset(x, type)

\method{loci.subset}{BayeScanData}(x, loci)

\method{loci.subset}{BayeScan}(x, loci)
}
\arguments{
\item{x}{\code{BayeScanData}, or \code{BayeScan}.}

\item{type}{\code{numeric} index of loci to subset. If \code{x} is a \code{BayeScan} object, then a \code{character} denoting the type of loci to subset. Valid arguments are 'adaptive', or 'neutral'.}
}
\value{
\code{BayeScanData}
}
\description{
This function returns a subset of loci in a \code{BayeScan} object.
}

