% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.multi.species.AttributeSpace.R
\name{make.multi.species.AttributeSpace}
\alias{make.multi.species.AttributeSpace}
\title{Make multi-species AttributeSpace object}
\usage{
make.multi.species.AttributeSpace(site.data, species.data)
}
\arguments{
\item{site.data}{\code{data.frame} with values for planning units in each attribute space.}

\item{species.data}{\code{data.frame} with binary occupancy data. Each column is for a differnet species. Each row is for a different planning unit.}
}
\value{
\code{AttributeSpace}
}
\description{
This function generates an \code{AttributeSpace} object using 
values in a \code{data.frame} and occupancy data in a second 
\code{data.frame}.
}
\details{
The demand points for each species are generated as the coordinates of the planning units
that the species is found in.
}
\seealso{
\code{\link[rapr]{AttributeSpace}}, \code{\link[rapr]{DemandPoints}}.
}

