% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.single.species.AttributeSpace.R
\name{make.single.species.AttributeSpace}
\alias{make.single.species.AttributeSpace}
\title{Make AttributeSpace using data for one species}
\usage{
make.single.species.AttributeSpace(site.data, species.data, n.species, spp.pos)
}
\arguments{
\item{site.data}{\code{data.frame} with coordinates for planning units.}

\item{species.data}{\code{data.frame} with coordinates of demand points for the species.}

\item{n.species}{\code{numeric} number of species to output.}

\item{spp.pos}{\code{numeric} index of this species.}
}
\value{
\code{AttributeSpace}.
}
\description{
This function creates an \code{AttributeSpace} object with multiple speciess. It
has a valid \code{DemandPoints} object for a single species,
and has dummy \code{DemandPoints} objects for the remaining species.
}
\seealso{
\code{\link[rapr]{AttributeSpace}}, \code{\link[rapr]{AttributeSpace}}.
}

