% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.targets.R
\name{make.targets}
\alias{make.targets}
\title{Make target data using attribute spaces}
\usage{
make.targets(species, environmental.space, geographic.space, adaptive.spaces,
  neutral.spaces, amount.target = 0.2, space.target = 0.2)
}
\arguments{
\item{species}{\code{character} speces names.}

\item{environmental.space}{\code{AttributeSpace} object with environmental space data.}

\item{geographic.space}{\code{AttributeSpace} object with geographic space data.}

\item{adaptive.spaces}{\code{list} of \code{AttributeSpace} objects with adaptive space data for species with loci undergoing adaptation.}

\item{neutral.spaces}{\code{list} of \code{AttributeSpace} objects with netrual space data for species with loci not undergoing adaptation.}
}
\description{
This function takes a set of species names and 
attribute spaces and generates a table with targets.
}
\details{
The \code{AttributeSpace} objects in the arguments to \code{adaptive.spaces} and \code{neutral.spaces} are special, such that the species they are not associated with will have negative weights.
}
\seealso{
\code{\link[rapr]{AttributeSpace}}.
}

