% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R, R/BayeScan.R
\name{mds}
\alias{mds}
\alias{mds.BayeScan}
\alias{mds.BayeScanData}
\title{Nonmetric multidimensional scaling for AFLP data}
\usage{
mds(x, metric, type, ...)

\method{mds}{BayeScanData}(x, metric = "gower", type = "all", ...)

\method{mds}{BayeScan}(x, metric, type, ...)
}
\arguments{
\item{x}{\code{BayeScanData} object.}

\item{metric}{\code{character} name of distance metric to use. Valid arguments are 'euclidean', 'manhattan', or 'gower'. Defaults to 'gower'.}

\item{type}{\code{character} type of loci to analyse. Valid arguments are 'all', 'adaptive', or 'neutral'. Defaults to 'all'.}

\item{...}{arguments passed to \code{link[vegan]{metaMDS}}.}
}
\value{
\code{\link[vegan]{metaMDS}} object.
}
\description{
This function performs nonmetric multidimensional scaling analysis on the loci in a \code{BayeScan} object.
}
\details{
The \code{link[cluster]{daisy}} function is used to calculate distances because it can accomodate missing values.
}
\examples{
# run BayeScan using low number of iterations
dat <- read.BayeScanData('int/Data_Meirmans_et_al_IntrabioDiv/Androsace_obtusifolia_AFLP.dat')
x <- run.BayeScan(x, threads=1, n=50, thin=1, nbp=10, pilot=10, burn=10)
z <- mds(x, metric='gower' k=2)
}
\seealso{
\code{link[cluster]{daisy}}, \code{link[vegan]{metaMDS}}.
}

