% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R, R/generics.R, R/BayeScanData.R, R/BayeScan.R
\name{n.pop.BayeScanResults}
\alias{n.pop}
\alias{n.pop.BayeScan}
\alias{n.pop.BayeScanData}
\alias{n.pop.BayeScanResults}
\title{Number of populations}
\usage{
\method{n.pop}{BayeScanResults}(x)

n.pop(x)

\method{n.pop}{BayeScanData}(x)

\method{n.pop}{BayeScan}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{integer}
}
\description{
This function returns the number of populations in a \code{BayeScan} object.
}

