% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R, R/generics.R, R/BayeScanData.R, R/BayeScan.R
\name{n.samples.BayeScanResults}
\alias{n.samples}
\alias{n.samples.BayeScan}
\alias{n.samples.BayeScanData}
\alias{n.samples.BayeScanResults}
\title{Number of samples}
\usage{
\method{n.samples}{BayeScanResults}(x)

n.samples(x)

\method{n.samples}{BayeScanData}(x)

\method{n.samples}{BayeScan}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{integer}
}
\description{
This function returns the number of samples in a \code{BayeScan} object.
}

