% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/BayeScanData.R
\name{pop.subset}
\alias{pop.subset}
\alias{pop.subset.BayeScanData}
\title{Subset populations}
\usage{
pop.subset(x, populations)

\method{pop.subset}{BayeScanData}(x, populations)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}

\item{populations}{\code{character} name of populations to subset.}
}
\value{
\code{BayeScanData}
}
\description{
This function returns a subset of populations in a \code{BayeScan} object.
}

