% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanData.R
\name{read.BayeScanData}
\alias{read.BayeScanData}
\title{Read FSTAT data for BayeScan}
\usage{
read.BayeScanData(x)
}
\arguments{
\item{x}{\code{character} file path with data}
}
\value{
\code{BayeScanData}.
}
\description{
This function reads FSTAT data.
}
\examples{
x <- read.BayeScanData('int/Data_Meirmans_et_al_IntrabioDiv/Androsace_obtusifolia_AFLP.dat')
}
\seealso{
\code{BayeScanData}.
}

