% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R
\name{read.BayeScanResults}
\alias{read.BayeScanResults}
\title{Read BayeScan results}
\usage{
read.BayeScanResults(file, dir, threshold = 0.95)
}
\arguments{
\item{file}{\code{character} file path of input file.}

\item{dir}{\code{dir} directory with output files.}

\item{threshold}{\code{numeric} threshold probability to classify loci as adaptive.}
}
\value{
\code{\link{BayeScanResults}}.
}
\description{
This function reads the results the BayeScan program.
}
\seealso{
\code{\link{BayeScanResults-class}}.
}

