% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScan.R
\name{run.BayeScan}
\alias{run.BayeScan}
\title{Run BayeScan}
\usage{
run.BayeScan(x, threads = 1, n = 5000, thin = 10, nbp = 20,
  pilot = 5000, burn = 50000, threshold = 0.95, dir = tempdir(),
  clean = TRUE)
}
\arguments{
\item{x}{\code{BayeScanData} object.}

\item{threads}{\code{numeric} number of threads used for computation. Defaults to 1.}

\item{n}{\code{numeric} number of outputted iterations. Defaults to 5000.}

\item{thin}{\code{numeric} thinning interval size. Defaults to 10.}

\item{nbp}{\code{numeric} number of pilot runs. Defaults to 20.}

\item{pilot}{\code{numeric} length of pilot runs. Defaults to 5000.}

\item{burn}{\code{numeric} burn-in length. Defaults to 50000.}

\item{threshold}{\code{numeric} threshold probability used to classify loci as adaptive. Defaults to 0.95.}

\item{dir}{\code{character} with directory to use for analysis.}

\item{clean}{\code{logical} should input and output files be deleted after analysis is finished?}
}
\description{
This function analysis data using BayeScan.
}
\examples{
# run BayeScan using low number of iterations
dat <- read.BayeScanData('int/Data_Meirmans_et_al_IntrabioDiv/Androsace_obtusifolia_AFLP.dat')
x <- run.BayeScan(x, threads=1, n=50, thin=1, nbp=10, pilot=10, burn=10)
}
\seealso{
\code{BayeScanData}, \code{BayeScanOpts}.
}

