% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R, R/generics.R, R/BayeScanData.R, R/BayeScan.R
\name{sample.pops.BayeScanResults}
\alias{sample.pops}
\alias{sample.pops.BayeScan}
\alias{sample.pops.BayeScanData}
\alias{sample.pops.BayeScanResults}
\title{Name of sample populations}
\usage{
\method{sample.pops}{BayeScanResults}(x)

sample.pops(x)

\method{sample.pops}{BayeScanData}(x)

\method{sample.pops}{BayeScan}(x)
}
\arguments{
\item{x}{\code{BayeScanData}, \code{BayeScan}.}
}
\value{
\code{character}
}
\description{
This function returns the population name that each sample belongs to in a \code{BayeScan} object.
}

