% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanResults.R, R/generics.R, R/BayeScanData.R, R/BayeScanOpts.R, R/BayeScan.R
\docType{methods}
\name{show,BayeScanResults-method}
\alias{show}
\alias{show,BayeScan-method}
\alias{show,BayeScanData-method}
\alias{show,BayeScanOpts-method}
\alias{show,BayeScanResults-method}
\title{Show objects}
\usage{
\S4method{show}{BayeScanResults}(object)

\S4method{show}{BayeScanData}(object)

\S4method{show}{BayeScanOpts}(object)

\S4method{show}{BayeScan}(object)
}
\arguments{
\item{object}{the object to show.}
}
\description{
This function shows objects.
}

