% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.prioritisation.R
\name{species.prioritisation}
\alias{species.prioritisation}
\title{Generate prioritisation using specific targets}
\usage{
species.prioritisation(x, amount.targets, env.surrogate.targets,
  geo.surrogate.targets, adaptive.genetic.targets, neutral.genetic.targets, ...)
}
\arguments{
\item{x}{\code{RapUnsolved} object.}

\item{amount.targets}{\code{numeric} amount-based targets to use.}

\item{env.surrogate.targets}{\code{numeric} environmental surrogate targets.}

\item{geo.surrogate.targets}{\code{numeric} geographic surrogate targets to use.}

\item{adaptive.genetic.targets}{\code{numeric} adaptive genetic targets.}

\item{neutral.genetic.targets}{\code{numeric} neutral genetic targets.}

\item{...}{arguments passed to solve.}
}
\value{
\code{RapSolved}
}
\description{
This function is a wrapper for updating targets in a \code{RapUnsolved} object and solving it.
}

