% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{spFortify}
\alias{spFortify}
\title{spFortify}
\usage{
spFortify(x, columns = names(x@data))
}
\arguments{
\item{x}{\code{\link[sp]{SpatialLinesDataFrame}}, \code{\link[sp]{SpatialPointsDataFrame}}, or \code{\link[sp]{SpatialPolygonsDataFrame}} object.}

\item{columns}{\code{character} with column names to include in returned object.}
}
\value{
\code{data.frame}
}
\description{
This function is a convienience function designed to prepare a \code{Spatial*DataFrame} object
for visualisation using ggplot2.
}
\seealso{
\code{\link[ggplot2]{fortify}}, \code{\link[sp]{SpatialLinesDataFrame}}, \code{\link[sp]{SpatialPointsDataFrame}} \code{\link[sp]{SpatialPolygonsDataFrame}}.
}

