% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgs1984}
\alias{wgs1984}
\title{WGS1984 CRS}
\format{An object of class \code{CRS} of length 1.}
\usage{
wgs1984
}
\description{
This object stores the coordinate reference system (CRS) for latitude/longitude data (WGS1984).
}
\examples{
wgs1984 <- CRS('+init=epsg:4326')
}
\seealso{
\code{\link[sp]{CRS}}
}
\keyword{datasets}

