% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeScanData.R
\name{write.BayeScanData}
\alias{write.BayeScanData}
\title{Write data for BayeScan}
\usage{
write.BayeScanData(x, file)
}
\arguments{
\item{x}{\code{BayeScanData} object.}

\item{file}{\code{character} file path to write data.}
}
\description{
This function writes data for BayeScan.
}
\details{
This code is based on the R code distributed with BayeScan (version 2.1) for converting binary FSTAT files to the BayeScan data format.
}
\examples{
x <- read.BayeScanData('int/Data_Meirmans_et_al_IntrabioDiv/Androsace_obtusifolia_AFLP.dat')
write.BayeScanData(x, tempfile(fileext='txt'))
}
\seealso{
\code{BayeScanData}.
}

