% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{europeEA}
\alias{europeEA}
\title{Europe Lambert Conformal Conic CRS}
\format{An object of class \code{CRS} of length 1.}
\usage{
europeEA
}
\description{
This object stores the coordinate reference system (CRS) for the Europe Lambert Conformal Conic projection.
}
\examples{
europeEA <- CRS('+init=esri:102014')
}
\seealso{
\code{\link[sp]{CRS}}
}
\keyword{datasets}

