% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.prioritisations.R
\name{species.prioritisation}
\alias{species.prioritisation}
\title{Generate prioritisation using specific targets}
\usage{
species.prioritisation(x, amount.targets, surrogate.targets, genetic.targets)
}
\arguments{
\item{x}{\code{RapUnsolved} object.}

\item{amount.targets}{\code{numeric} amount-based targets to use.}

\item{surrogate.targets}{\code{numeric} surrogate-based targets to use.}

\item{genetic.targets}{\code{numeric} genetic-based targets to use.}

\item{...}{arguments passed to solve.}
}
\value{
\code{RapSolved}
}
\description{
This function is a wrapper for updating targets in a \code{RapUnsolved} object and solving it.
}

